pyproject.toml
[project]
name = "wild-goose-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "dotenv>=0.9.9",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.6",
    "loguru>=0.7.3",
    "markdown>=3.5",
    "openai>=2.14.0",
    "playwright>=1.58.0",
    "PySide6>=6.6",
    "python-frontmatter>=1.1.0",
    "qasync>=0.27.1",
    "qtawesome>=1.3",
    "rich>=14.2.0",
]

[[tool.uv.index]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff.format]
line-ending = "lf"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
